client: manage remote peer disconnection
Before this commit, if the connection was closed by the remote
peer (e.g. the server was terminated), `ClientProtocol` didn't properly
communicate the end of the protocol to the observing objects, putting
the app in an inconsistent state (well almost: everything still worked,
just wasn't aware that the connection was closed).